+++ changelog.md
@@ -43,6 +43,7 @@
 * [1879](https://github.com/zeta-chain/node/pull/1879) - full coverage for messages in types packages
 * [1899](https://github.com/zeta-chain/node/pull/1899) - add empty test files so packages are included in coverage
 * [1903](https://github.com/zeta-chain/node/pull/1903) - common package tests
+* [1941](https://github.com/zeta-chain/node/pull/1941) - add unit tests for zetabridge package
 
 ### Fixes